+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - PDO: Preserve the exact value of double (bug #1214)
 - PostgreSQL, MS SQL, Oracle: Do not highlight a valid date or time value as too long
 - PostgreSQL: Alter the trigger function from table structure
+- PostgreSQL: Do not highlight an array value as too long
 - MS SQL: Export and alter the precision of datetime2, time and datetimeoffset
 - MS SQL: Keep fractional seconds and the time zone offset when saving a row with the sqlsrv extension
 - MS SQL Dblib: Display uniqueidentifier (bug #1339)